jxl-color: Detect simple ICC profiles
